pyproject.toml
[project]
name = "note-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "python-dotenv>=1.2.2",
    "requests>=2.32.3",
    "faiss-cpu>=1.7.4",
    "pypdf>=4.0.0",
    "sentence-transformers>=2.5.0",
    "numpy>=1.24.0",
    "langgraph>=1.0.10",
]

[tool.setuptools]
py-modules = ["main"]
